web/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
uWSGI==2.0.18